# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "JD任务平台后端API服务 - Go语言实现版本",
        "title": "JD任务平台 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@jdtask.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api",
    "paths": {
...
        "/tasks/apikey/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量创建任务，最多100个",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "批量创建任务",
                "parameters": [
                    {
                        "description": "批量任务信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BatchCreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tasks/statistics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取任务统计数据，管理员可指定user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "获取任务统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID（仅管理员）",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/tasks/types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有可用的任务类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "获取任务类型列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的任务类型（仅管理员）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "创建任务类型",
                "parameters": [
                    {
                        "description": "任务类型信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tasks/types/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新任务类型信息（仅管理员）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "更新任务类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "任务类型信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTaskTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取消等待中的任务并退款京豆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "取消任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tasks/{id}/priority": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改任务优先级（仅管理员）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务模块"
                ],
                "summary": "修改任务优先级",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "优先级",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "priority": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BatchCreateTaskRequest": {
            "type": "object",
            "required": [
                "tasks"
            ],
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateTaskRequest"
                    }
                }
            }
        },
...
        "models.CreateTaskRequest": {
            "type": "object",
            "required": [
                "consume_jingdou",
                "execute_count",
                "sku",
                "start_time",
                "task_type"
            ],
            "properties": {
                "consume_jingdou": {
                    "type": "integer",
                    "example": 100
                },
                "execute_count": {
                    "type": "integer",
                    "example": 10
                },
                "keyword": {
                    "type": "string",
                    "example": "手机"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "remark": {
                    "type": "string",
                    "example": "测试任务"
                },
                "shop_name": {
                    "type": "string",
                    "example": "京东自营店"
                },
                "sku": {
                    "type": "string",
                    "example": "100001234567"
                },
                "start_time": {
                    "type": "string",
                    "example": "2023-12-01T10:00:00Z"
                },
                "task_type": {
                    "type": "string",
                    "example": "search_order"
                }
            }
        },
        "models.CreateTaskTypeRequest": {
            "type": "object",
            "required": [
                "jingdou_price",
                "type_code",
                "type_name"
            ],
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "jingdou_price": {
                    "type": "integer",
                    "example": 2
                },
                "type_code": {
                    "type": "string",
                    "example": "add_to_cart"
                },
                "type_name": {
                    "type": "string",
                    "example": "加购"
                }
            }
        },
...
        "models.UpdateTaskTypeRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": false
                },
                "jingdou_price": {
                    "type": "integer",
                    "example": 3
                },
                "type_name": {
                    "type": "string",
                    "example": "新名称"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "操作成功"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for authentication",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}